Fix userspace warnings.

+++ us/twix/syscalls.h
@@ -90,9 +90,9 @@ long linux_sys_arch_prctl(int code, unsigned long addr);
 
 long linux_sys_getrandom(char *, size_t, unsigned int);
 
-int linux_sys_dup(int oldfd);
-int linux_sys_dup2(int oldfd, int newfd);
-int linux_sys_dup3(int oldfd, int newfd, int flags);
+long linux_sys_dup(int oldfd);
+long linux_sys_dup2(int oldfd, int newfd);
+long linux_sys_dup3(int oldfd, int newfd, int flags);
 long linux_sys_open(const char *path, int flags, int mode);
 long linux_sys_close(int fd);
 long linux_sys_lseek(int fd, off_t off, int whence);
@@ -119,7 +119,7 @@ long linux_sys_set_tid_address();
 long linux_sys_fork(struct twix_register_frame *frame);
 long linux_sys_wait4(long pid, int *wstatus, int options, struct rusage *rusage);
 
-int linux_sys_mkdir(char *path, int mode);
+long linux_sys_mkdir(char *path, int mode);
 long linux_sys_getegid(void);
 long linux_sys_geteuid(void);
 long linux_sys_getgid(void);
